MOVE: Core infrastructure files/queues to separate packages
