+++ src/lib/kotlin/slatekit-core/src/main/kotlin/slatekit/core/docs/CloudDoc.kt
@@ -1,0 +1,11 @@
+package slatekit.core.docs
+
+interface CloudDoc<TEntity, TPartition, TCluster> {
+    val partition:TPartition
+
+    val cluster:TCluster
+
+    val fields:Map<String, Any?>
+
+    val source:Any
+}
