+++ src/lib/kotlin/slatekit/src/main/kotlin/slatekit/SlateKitServices.kt
@@ -5,8 +5,8 @@ import slatekit.apis.core.Api
 import slatekit.cloud.aws.AwsCloudFiles
 import slatekit.cloud.aws.AwsCloudQueue
 import slatekit.common.queues.QueueStringConverter
-import slatekit.core.cloud.CloudFiles
-import slatekit.core.cloud.CloudQueue
+import slatekit.core.files.CloudFiles
+import slatekit.core.queues.CloudQueue
 import slatekit.notifications.email.EmailService
 import slatekit.notifications.email.EmailServiceSendGrid
 import slatekit.notifications.sms.SmsService
